backend/requirements.txt
fastapi
uvicorn
pandas
numpy
faiss-cpu
pillow
python-multipart
groq
pydantic
python-dotenv
torch
transformers